fix: ajuste no fehcmaneto da conta, influenciada pela troca de Float -> BigDecimal no MedicalInsurance

+++ src/main/java/com/hospital/servicesprovidedclosure/service/ProvidedClosureService.java
@@ -48,7 +48,7 @@ public class ProvidedClosureService {
             throw new RuntimeException("Paciente nao possui convenio");
         }
 
-        BigDecimal coverageRate = BigDecimal.valueOf(medicalInsurance.getCoverageRate());
+        BigDecimal coverageRate = medicalInsurance.getCoverageRate();
 
         List<ServicesProvided> despesas = servicesProvidedRepository.findByAdmission_Id(admissionId);
